Prevent NPE if shared library isn't in SDRROOT
If a softpkg has a dependency on another softpkg, and that softpkg isn't installed in the SDRROOT,
an NPE was thrown while our external settings provider for CDT was creating a list of include paths.
Now, the referenced (and non-installed) softpkg will cause an include entry to be created (which
obviously won't be present), and the settings provider won't attempt to recursively provide children
for the referenced softpkg.