IDE-1353 Make the error parser recognize newer autotools messages
A newer version of autotools in use on Ubuntu 14.04 LTS is causing issues because the format
of one of the messages it emits has changed. Normally CDT believes this message to be an error,
but we filter it out with a custom error filter. The format of the message changed by one
character - from a back-tick to an apostrophe. This commit makes our regex recognize both.